interface implementation

+++ Com.Danliris.Service.Merchandise.Lib/Models/RO_Garment_SizeBreakdown.cs
@@ -11,7 +11,7 @@ namespace Com.Danliris.Service.Merchandiser.Lib.Models
         public int RO_GarmentId { get; set; }
         public virtual RO_Garment RO_Garment { get; set; }
         public string Code { get; set; }
-        public int ColorId { get; set; }
+        public long ColorId { get; set; }
         public string ColorName { get; set; }
         public ICollection<RO_Garment_SizeBreakdown_Detail> RO_Garment_SizeBreakdown_Details { get; set; }
         public int Total { get; set; }
